Switch back to double precision

+++ voxblox/include/voxblox/core/common.h
@@ -38,7 +38,7 @@ inline std::shared_ptr<Type> aligned_shared(Arguments&&... arguments) {
 }
 
 // Types.
-typedef float FloatingPoint;
+typedef double FloatingPoint;
 typedef int IndexElement;
 
 typedef Eigen::Matrix<FloatingPoint, 3, 1> Point;
